requirements_test.txt
amqp==2.6.0
anyconfig==0.9.11
anyjson==0.3.3
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
base58==2.0.0
billiard==3.6.3.0
bitarray==1.2.1
bitcoin==1.1.42
bleach==1.5.0
cached-property==1.5.1
celery==4.4.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
coincurve==13.0.0
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.10.1
descartes==1.1.0
Django==3.0.6
django-ipfs-storage==0.0.4
django-web3-auth==0.1.0
enum34==1.1.10
eth-abi==0.5.0
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.1.0b4
eth-rlp==0.1.2
eth-tester==0.1.0b11
eth-testrpc==1.3.5
eth-typing==2.2.1
eth-utils==0.8.1
ethereum==2.3.1
Fiona==1.8.13.post1
future==0.18.2
geopandas==0.3.0
hexbytes==0.2.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.0
ipfs-api==0.2.3
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
json-rpc==1.13.0
jsonschema==3.2.0
kombu==4.6.10
lru-dict==1.1.6
Markdown==3.2.2
matplotlib==2.1.2
more-itertools==8.3.0
multiaddr==0.0.9
munch==2.5.0
mypy-extensions==0.4.3
netaddr==0.7.19
numpy==1.18.4
packaging==20.3
pandas==0.22.0
parsimonious==0.8.1
pathtools==0.1.2
pbkdf2==1.3
pkg-resources==0.0.0
pluggy==0.13.1
populus==2.2.0
protobuf==3.12.0
py==1.8.1
py-ecc==4.0.0
py-geth==2.3.0
py-solc==3.2.0
pycparser==2.20
pycryptodome==3.9.7
pyethash==0.1.27
pylru==1.2.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
pysha3==1.0.2
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==2.10.3
repoze.lru==0.7
requests==2.23.0
rlp==0.6.0
scrypt==0.8.13
secp256k1==0.13.2
semantic-version==2.8.5
Shapely==1.7.0
six==1.14.0
sqlparse==0.3.1
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0
toolz==0.10.0
toposort==1.5
typing-extensions==3.7.4.2
urllib3==1.25.9
varint==1.0.2
vine==1.3.0
watchdog==0.10.2
wcwidth==0.1.9
web3==3.16.5
websockets==8.1
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==3.1.0
